refactor: consolidate effect imports across authentication modules and remove unused endpointResponses file
