+++ packages/studiocms/src/lib/auth/password.ts
@@ -2,8 +2,7 @@ import crypto from 'node:crypto';
 import { FetchHttpClient, HttpClient } from '@effect/platform';
 import { sha1 } from '@oslojs/crypto/sha1';
 import { encodeHexLowerCase } from '@oslojs/encoding';
-import { Data, Effect } from 'effect';
-import { genLogger, pipeLogger } from '../effects/index.js';
+import { Data, Effect, genLogger, pipeLogger } from '../../effect.js';
 import { Scrypt } from './utils/scrypt.js';
 import { CheckIfUnsafe, type CheckIfUnsafeError } from './utils/unsafeCheck.js';
