+++ packages/studiocms/src/lib/auth/user.ts
@@ -2,8 +2,7 @@ import { Notifications } from 'studiocms:notifier';
 import { SDKCoreJs as sdk } from 'studiocms:sdk';
 import type { CombinedUserData, tsUsersInsert } from 'studiocms:sdk/types';
 import type { APIContext, AstroGlobal } from 'astro';
-import { Data, Effect, pipe } from 'effect';
-import { genLogger, pipeLogger } from '../effects/index.js';
+import { Data, Effect, genLogger, pipe, pipeLogger } from '../../effect.js';
 import { Password } from './password.js';
 import { Session } from './session.js';
 import type { UserSessionData } from './types.js';
