+++ packages/studiocms/src/lib/auth/utils/scrypt.ts
@@ -1,7 +1,15 @@
 import { CMS_ENCRYPTION_KEY } from 'astro:env/server';
 import { type ScryptOptions, scrypt } from 'node:crypto';
-import { Brand, Context, Data, Effect, Layer } from 'effect';
-import { errorTap, genLogger, pipeLogger } from '../../effects/index.js';
+import {
+	Brand,
+	Context,
+	Data,
+	Effect,
+	errorTap,
+	genLogger,
+	Layer,
+	pipeLogger,
+} from '../../../effect.js';
 
 const parsedN = Number.parseInt(process.env.SCRYPT_N ?? '', 10);
 const SCRYPT_N = Number.isFinite(parsedN) ? Math.max(16384, parsedN) : 16384;
